requirements.txt

Pillow
django-crispy-forms
django-allauth
redis-server
python-decouple

